backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
websockets
passlib[bcrypt]
python-jose[cryptography]
pyjwt
python-multipart
paho-mqtt
